+++ src/main/java/org/sopt/repository/PostRepository.java
@@ -1,37 +1,15 @@
 package org.sopt.repository;
 
 import org.sopt.domain.Post;
-import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-@Repository
-public class PostRepository {
-    private final List<Post> postList = new ArrayList<>();
-    private Long nextId = 1L;
+public interface PostRepository {
 
-    public Post save(Post post) {
-        postList.add(post);
-        return post;
-    }
-
-    public List<Post> findAll() {
-        return postList;
-    }
-
-    public Optional<Post> findById(Long id) {
-        return postList.stream()
-                .filter(p -> p.getId().equals(id))
-                .findFirst();
-    }
-
-    public boolean deleteById(Long id) {
-        return postList.removeIf(p -> p.getId().equals(id));
-    }
-
-    public Long generateId() {
-        return nextId++;
-    }
+    Post save(Post post);
+    List<Post> findAll();
+    Optional<Post> findById(Long id);
+    boolean deleteById(Long id);
+    Long generateId();
 }
